[Feat]: Setting 기능 수정
